AB#61 Create service layer test suite (#8)
* create personality model

* disable pipelines on non-PR associated pushes

* fix typo

* initial test structure

* generate basic unit tests for each src file

* fix some malformed test errors

* remove hibernate testing

* no need for any hibernate config

* getters and setters

There may be items here that should not have public getters and setters.
Passwords, for instance, are suspect. However, for the sake of being
able to test, for the time being we're adding them. This may merit
further refinement if capacity allows.

* remove inapplicable tests and improve logic of testCreatedBeforeUpdated and set access from public to default

No need to test that lobby.getId is not null, since ID is supplied by the database.

Rewrote logic for testCreatedBeforeUpdated to allow the two times to be equal and still pass

* make tests clean and passable

Now that getters and setters exist, it's easier to see what tests need in order to meaningfully test

* ignore coverage report

* remove hibernate dependencies and reformat, including using jwtt version property

* fix camelcase

* fix camelcase

* fix getGameById -> getPartyById and fix camelcase

* handle potential get user failures

* improve springboot config annotations

* fix springboot config tag and make tests make more sense with given constructor

* fix issues with repository tests

This required setting up an h2 internal database to mimic database
functionality without actually connecting to the database.

There is an issue with the hardcoded query in PartyRepository, and once
that is fixed we can uncomment the test in PartyRepositoryTest